feat: form validation for registration/login
-Add/configure login & registration page. -Add JSON file to apply defaults to form label and inputs. -Add form validation to  username, email, password, and password confirmation inputs.
