+++ client/src/components/forms/styles/FormInputs.module.css
@@ -1,0 +1,11 @@
+.error {
+  display: none;
+}
+
+input:invalid[focused="true"] {
+  border: 1px solid red;
+}
+
+input:invalid[focused="true"] + .error {
+  display: block;
+}
