+++ client/src/pages/Registration.tsx
@@ -1,12 +1,12 @@
 import LoginMenu from "../components/forms/SubmissionForm";
 import formData from "../local-json/formInputs.json"; //Contains input & label defaults for form
 
-function Login() {
+function Registration() {
   return (
     <div className="flex justify-center my-20">
-      <LoginMenu formData={formData.login} />
+      <LoginMenu formData={formData.registration} />
     </div>
   );
 }
 
-export default Login;
+export default Registration;
